TOURNAMENT HANDLING
- updated DAO to contains also tournament access. There is a gotcha:
Partecipations is not a entity class, hence it is not stored inside the
maps in DAO. However, we need to be reactive to those changes. There are
2 phases to ensure that: first of all when we create either a team or a
tournament we add ot the getPssrtecipation() observablelist a listener;
such listener will add/remove partecipations inside the db when the
observable list changes: in this way we are reactive when the user
modify an existing main concept. The seconds change involves the fetch
from the database: everytime a new team/tournament is created, we create
the references inside the model in the memory, without affecting the DB